requirements.txt
streamlit
openai
PyPDF2
tiktoken
langchain
pandas
python-docx
openpyxl



